move pages to separate directory
