+++ frontend/packingplans/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {HomepageComponent} from "./components/homepage/homepage.component";
-import {LocationPageComponent} from "./components/location-page/location-page.component";
+import {HomepageComponent} from "./pages/homepage/homepage.component";
+import {LocationPageComponent} from "./pages/location-page/location-page.component";
 
 const routes: Routes = [
   {path: "home", title: "PackingPlans - Home", component: HomepageComponent},
