+++ frontend/packingplans/src/app/pages/location-page/location-page.component.ts
